he separado la funcionalidad de mostrar u ocultar los formularios de register y form. Lo he hecho mas modular
