+++ views/_login-forms.ejs
@@ -15,7 +15,7 @@
 
     <button id="login-btn">Login</button>
 
-    <i class="fa-solid fa-x" id="close-login-form-btn"></i>
+    <i class="fa-solid fa-x" id="close-login-form-btn" data-modal-target="login-form" data-modal-action="close"></i>
 </form>
 
 <!-- method="POST" action="/user/login" -->
@@ -44,5 +44,5 @@
 
     <button id="register-btn">Register</button>
 
-    <i class="fa-solid fa-x" id="close-register-form-btn"></i>
+    <i class="fa-solid fa-x" id="close-register-form-btn" data-modal-target="register-form" data-modal-action="close"></i>
 </form>
