+++ views/index.ejs
@@ -64,10 +64,10 @@
     <div id="overlay"></div>
 
     <% if (user) { %>
-        <script src="/js/header.js"></script>
+        <script src="/js/header/header.js" type="module"></script>
     <% } else { %>
         <%- include('./_login-forms.ejs') %>
-        <script src="/js/unlogged-users-header.js"></script>
+        <script src="/js/unlogged-users-header/unlogged-users-header.js" type="module"></script>
     <% } %> 
 
     <script src="/js/index.js" type="module"></script>
