+++ views/unlogged-users-header.ejs
@@ -1,8 +1,8 @@
 <header>
     <h2>THE BLOG</h2>
     <div id="header-btns">
-        <button id="show-login-form-btn">Login</button>
+        <button id="show-login-form-btn" data-modal-target="login-form" data-modal-action="open">Login</button>
 
-        <button id="show-register-form-btn">Register</button>
+        <button id="show-register-form-btn" data-modal-target="register-form" data-modal-action="open">Register</button>
     </div>
 </header>
